This is hidden in v2.55.0-rc0's own CI because of an omission in
5ba82911bccc (ci: enable EXPENSIVE for contributor builds, 2026-05-11)
which fails to enable EXPENSIVE tests for tags.

Due to 7d78d5fc1a91 (ci: skip GitHub workflow runs for already-tested
commits/trees, 2020-10-08), the CI of `master` is now also mistakenly
green because it reuses the tag's CI run to prove that it's solid.

This is an evil merge by necessity because `survey.c` needs to adapt to
the changed function signatures.

+++ builtin/survey.c
@@ -636,7 +636,7 @@ static void survey_phase_refs(struct survey_context *ctx)
 
 	ctx->report.refs.refs_nr = ctx->ref_array.nr;
 	for (int i = 0; i < ctx->ref_array.nr; i++) {
-		unsigned long size;
+		size_t size;
 		struct ref_array_item *item = ctx->ref_array.items[i];
 
 		switch (item->kind) {
@@ -703,7 +703,7 @@ static void increment_totals(struct survey_context *ctx,
 	for (size_t i = 0; i < oids->nr; i++) {
 		struct object_info oi = OBJECT_INFO_INIT;
 		unsigned oi_flags = OBJECT_INFO_FOR_PREFETCH;
-		unsigned long object_length = 0;
+		size_t object_length = 0;
 		off_t disk_sizep = 0;
 		enum object_type type;
